Add TextRun

+++ nfm-world/driverinterface/IFontMetrics.cs
@@ -2,6 +2,6 @@ namespace NFMWorld.DriverInterface;
 
 public interface IFontMetrics
 {
-    public int StringWidth(string astring);
+    public float StringWidth(string astring);
     public int Height(string astring);
 }
